+++ ios/flutter_techlabs_library/Sources/flutter_techlabs_library/FlutterTechlabsLibraryPlugin.swift
@@ -1,5 +1,6 @@
 import Flutter
 import UIKit
+import ios_techlabs_library
 
 public class FlutterTechlabsLibraryPlugin: NSObject, FlutterPlugin {